refactor(menu): split runtime wiring out so menu.rs is 100% testable
`menu.rs` previously mixed pure data and pure functions (MenuAction,
menu_action_for_id, dispatch_menu_action) with Tauri-runtime-bound
helpers (build_menu, apply_menu_state, handle_menu_event) that can't be
exercised from MockRuntime. The runtime half was dragging patch
coverage down on every menu change.

- Move build_menu / apply_menu_state / set_menu_state / handle_menu_event
  / MenuItemRefs / MenuStateFlags to `src-tauri/src/menu_runtime.rs`,
  re-using `menu_action_for_id` and `dispatch_menu_action` from menu.rs.
- Keep all the testable plumbing in `menu.rs` along with its dedicated
  tests (9 covering every match arm + every dispatch arm).
- Add `src-tauri/src/menu_runtime.rs` to `codecov.yml` `ignore:` so
  coverage credit is awarded to the testable surface only.
- Wire `lib.rs` up to the new module path; no behaviour change.
